+++ test/bdd/step/i_do_not_see_the_text.dart
@@ -1,0 +1,5 @@
+import 'package:flutter_test/flutter_test.dart';
+
+Future<void> iDoNotSeeTheText(WidgetTester tester, String label) async {
+  expect(find.text(label), findsNothing);
+}